Don't raise an exception at module load time
Most of the functions which fetch constants from the C code don't
fail... except the one that opens a null logger. Instead we make this
value "lazy" so it is not evaluated until we need it (by the Lazy.force)
which is a more appropriate time to raise the exception.

Signed-off-by: David Scott <dave.scott@docker.com>